update:更新AuthAttributeView位置 # Reviewed, transaction id: 36606
