+++ saas/backend/api/authorization/views/__init__.py
@@ -8,8 +8,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from .action import AuthAttributeView
-from .resource import AuthBatchInstanceView, AuthBatchPathView, AuthInstanceView, AuthPathView
+from .resource import AuthAttributeView, AuthBatchInstanceView, AuthBatchPathView, AuthInstanceView, AuthPathView
 from .resource_creator_action import (
     BatchResourceCreatorActionView,
     ResourceCreatorActionAttributeView,
